Prevent creating host container network

+++ src/Agent.Worker/ContainerOperationProvider.cs
@@ -35,7 +35,8 @@ namespace Microsoft.VisualStudio.Services.Agent.Worker
         {
             base.Initialize(hostContext);
             _dockerManger = HostContext.GetService<IDockerCommandManager>();
-            _containerNetwork = $"vsts_network_{Guid.NewGuid().ToString("N")}";
+            var useHostNetwork = AgentKnobs.DockerNetworkCreateDriver.GetValue(hostContext).AsString() == "host";
+            _containerNetwork = useHostNetwork ? "host" : $"vsts_network_{Guid.NewGuid().ToString("N")}";
         }
 
         public async Task StartContainersAsync(IExecutionContext executionContext, object data)
@@ -104,6 +105,7 @@ namespace Microsoft.VisualStudio.Services.Agent.Worker
             // Create local docker network for this job to avoid port conflict when multiple agents run on same machine.
             // All containers within a job join the same network
             await CreateContainerNetworkAsync(executionContext, _containerNetwork);
+
             containers.ForEach(container => container.ContainerNetwork = _containerNetwork);
 
             foreach (var container in containers)
@@ -138,6 +140,7 @@ namespace Microsoft.VisualStudio.Services.Agent.Worker
             {
                 await StopContainerAsync(executionContext, container);
             }
+
             // Remove the container network
             await RemoveContainerNetworkAsync(executionContext, _containerNetwork);
         }
@@ -676,11 +679,16 @@ namespace Microsoft.VisualStudio.Services.Agent.Worker
         {
             Trace.Entering();
             ArgUtil.NotNull(executionContext, nameof(executionContext));
-            int networkExitCode = await _dockerManger.DockerNetworkCreate(executionContext, network);
-            if (networkExitCode != 0)
+            
+            if (network != "host")
             {
-                throw new InvalidOperationException($"Docker network create failed with exit code {networkExitCode}");
+                int networkExitCode = await _dockerManger.DockerNetworkCreate(executionContext, network);
+                if (networkExitCode != 0)
+                {
+                    throw new InvalidOperationException($"Docker network create failed with exit code {networkExitCode}");
+                }
             }
+
             // Expose docker network to env
             executionContext.Variables.Set(Constants.Variables.Agent.ContainerNetwork, network);
         }
@@ -691,13 +699,17 @@ namespace Microsoft.VisualStudio.Services.Agent.Worker
             ArgUtil.NotNull(executionContext, nameof(executionContext));
             ArgUtil.NotNull(network, nameof(network));
 
-            executionContext.Output($"Remove container network: {network}");
-
-            int removeExitCode = await _dockerManger.DockerNetworkRemove(executionContext, network);
-            if (removeExitCode != 0)
+            if (network != "host")
             {
-                executionContext.Warning($"Docker network rm failed with exit code {removeExitCode}");
+                executionContext.Output($"Remove container network: {network}");
+
+                int removeExitCode = await _dockerManger.DockerNetworkRemove(executionContext, network);
+                if (removeExitCode != 0)
+                {
+                    executionContext.Warning($"Docker network rm failed with exit code {removeExitCode}");
+                }
             }
+
             // Remove docker network from env
             executionContext.Variables.Set(Constants.Variables.Agent.ContainerNetwork, null);
         }
